requirements.txt
scrapy
django==3.2.18
djangorestframework
django-pdb
django-cors-headers
psycopg2-binary
requests
django-cors-middleware
validators
django-admin-rangefilter
django-extensions
pyOpenSSL
drf-yasg
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
python-magic
urllib3
python-dateutil
python-dotenv
Werkzeug
pandas==1.1.5
opencv_python==4.5.4.58
numpy==1.19.3
munkres==1.1.4
tqdm==4.62.3
# mediapipe==0.8.3.1
absl-py==1.0.0
Pillow==8.4.0
# deepface
gunicorn
pyfcm
django-safedelete
django-pushy
django-rq
drf-api-logger>=1.0.10,<1.1.0
django-auditlog==2.0.0
django-crontab
#django-rest-swagger
websocket-client
websockets
pqdm
django-mailbox
django-celery
django-celery-beat
django-celery-results
django-jsonfield
django-redis
celery
gunicorn
requests 
redis
gevent
google-cloud-vision
google-api-python-client
















